broker-extraction-api/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
openpyxl==3.1.2
pillow==10.1.0
pypdf==3.17.4
paddleocr==3.2.0
uvicorn==0.24.0.post1
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
ultralytics==8.3.204

